core: pxe: additional work on the lwip port
Additional work on the lwip port.  With this code, we can get pretty
far before having problems.

This moves malloc into the core directory, mostly so we can guard the
malloc stuff with semaphores.  I suspect we're going to have to have
at least stub semaphores available to the library, since other things
are going to need to be made thread-safe, too.

This checkin contains debugging printfs and considerable amounts of
placeholder code, that needs to be fixed.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>